Fix progress token numeric type

+++ src/main/java/com/amannmalik/mcp/util/ProgressCodec.java
@@ -34,7 +34,7 @@ public final class ProgressCodec {
     private static ProgressToken toToken(JsonValue value) {
         return switch (value.getValueType()) {
             case STRING -> new ProgressToken.StringToken(((JsonString) value).getString());
-            case NUMBER -> new ProgressToken.NumericToken(((JsonNumber) value).longValue());
+            case NUMBER -> new ProgressToken.NumericToken(((JsonNumber) value).doubleValue());
             default -> throw new IllegalArgumentException("Invalid token type");
         };
     }
